environment/docker/requirements.txt
docker
PyYaml
kubernetes
click
lazy-object-proxy
colorful
tqdm
